AccName: Improve checks for redundant content/loops and hidden content
Check for hidden content moved from the beginning of the algorithm to the text-from-contents step; this changes the numbering of steps in the main algorithm, requiring corresponding updates to cross-references.

New details, including notes and examples regarding loops and redundant content.

Clears up definitions of hidden content in all specs.

Clears up definitions of current value for embedded widgets, in the core and HTML specs.

Adds check for the display property in CSS pseudo-elements.